pyproject.toml
[tool.poetry]
name = "poetry init"
version = "0.1.0"
description = ""
authors = ["Vil <Vil.king.li@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
psycopg2-binary = "^2.9.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
